Reservations on other dates
I made changes in the code so that the user does not have a reservation only for today's date, but can make a reservation on other dates as well.
Probably, the next step is to make it so that users who want to reserve a table can reserve the table and restaurant with the same name and authentication that they entered for registration.